Non windows support

+++ Modules.Tests/Google.Tests.ps1
@@ -7,8 +7,8 @@ Import-Module -Name $PSScriptRoot\..\Modules\IntelliTect.Google -Force
 Function Get-TestCredential {
     $credentialName = "IntelliTect.Google.Tests"
 
-    [PSCredential]$credential=Get-CredentialManagerCredential $credentialName -ErrorAction SilentlyContinue
-    if($credential -eq $null) {
+    [PSCredential]$credential = Get-CredentialManagerCredential $credentialName -ErrorAction SilentlyContinue
+    if ($credential -eq $null) {
         # Future versions (post 3.0) of Pester include 'Set-TestInconclusive' that should be used here.
         # For now, we just throw an error to discontinue the test while providing instructions on how to correctly configure the test.
         Set-TestInconclusive -Message "Couldn't find a credential named $credentialName. Add it using 'Set-CredentialManagerCredential $credentialName'. You will be prompted for a Google email account and password, which will be stored in the Windows credential manager."
@@ -17,32 +17,46 @@ Function Get-TestCredential {
 }
 
 Describe "Get-GoogleSessionVariable" {
-    It "Gets a WebRequestSession with Google session cookies" {
-        $credential = Get-TestCredential
+    if ($IsWindows) {
+        It "Gets a WebRequestSession with Google session cookies" {
+            $credential = Get-TestCredential
 
-        $session = Get-GoogleSession $credential
-        $sidCookie = ($session.Cookies.GetCookies("https://www.google.com") | where {$_.Name -eq "SID"}).Value
+            $session = Get-GoogleSession $credential
+            $sidCookie = ($session.Cookies.GetCookies("https://www.google.com") | where {$_.Name -eq "SID"}).Value
 
-        # Check that the SID cookie is at least 50 characters.
-        # One I used while testing was 71 characters, so 50 should be a safe minimum.
-        $sidCookie | Should Match ".{50,}"
+            # Check that the SID cookie is at least 50 characters.
+            # One I used while testing was 71 characters, so 50 should be a safe minimum.
+            $sidCookie | Should Match ".{50,}"
+        }
+    }
+    else {
+        It "Verify throws on non-Windows platforms" {
+            { Get-TestCredential } | Should -Throw 'This cmdlet is not supported on non-Windows operating systems.'
+        }
     }
 }
 
 Describe "Get-GoogleLocationHistoryKmlFile" {
-    It "Gets a kml file" {
-        $credential = Get-TestCredential
+    if ($IsWindows) {
+        It "Gets a kml file" {
+            $credential = Get-TestCredential
 
-        $session = Get-GoogleSession $credential
+            $session = Get-GoogleSession $credential
 
-        $outFile = Get-TempItemPath
-        $request = Get-GoogleLocationHistoryKmlFile $session 2016-05-01 -outFile $outFile
+            $outFile = Get-TempItemPath
+            $request = Get-GoogleLocationHistoryKmlFile $session 2016-05-01 -outFile $outFile
 
-        $outFile | Should Contain "http://www.opengis.net/kml/"
+            $outFile | Should Contain "http://www.opengis.net/kml/"
 
-        # cleanup
-        Get-Item $outFile | Remove-Item
-        Test-Path $outFile | Should Be $false
+            # cleanup
+            Get-Item $outFile | Remove-Item
+            Test-Path $outFile | Should Be $false
+        }
+    }
+    else {
+        It "Verify throws on non-Windows platforms" {
+            { Get-TestCredential } | Should -Throw 'This cmdlet is not supported on non-Windows operating systems.'
+        }
     }
 }
